feat: create new hook useNetworkProvider
